Refactoring footer, and remove globals and _mobile scss
