+++ components/Copyright.tsx
@@ -1,8 +1,9 @@
-import styles from "../styles/globals.module.scss";
+import styles from "../styles/copyright.module.scss";
 
 export default function Copyright() {
-    return (
-      <div className={styles.copyright}>©2020 <a href="/">resize.fm</a></div>
-    );
-  }
-  
+  return (
+    <div className={styles.copyright}>
+      ©2020 <a href="/">resize.fm</a>
+    </div>
+  );
+}
