--- styles/_mobile.scss
@@ -1,5 +1,0 @@
-@media screen and (max-width: 480px) {
-  .copyright {
-    font-size: 12px;
-  }
-}
