+++ styles/episode.module.scss
@@ -58,3 +58,8 @@
     }
   }
 }
+
+.footer {
+  background: $color-palegray2;
+  padding: 32px 0;
+}
