+++ styles/globals.scss
@@ -18,6 +18,10 @@ body {
   background-size: 1200px auto, 1px auto;
   background-repeat: no-repeat, repeat-x;
   text-align: center;
+
+  @include mobile {
+    background-size: 800px auto, 1px auto;
+  }
 }
 
 a {
@@ -26,11 +30,5 @@ a {
 
   &:hover {
     text-decoration: underline;
-  }
-}
-
-@media screen and (max-width: 480px) {
-  body {
-    background-size: 800px auto, 1px auto;
   }
 }
